exp/Inappropriate_tool_use/code/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncio==4.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
fastapi==0.116.1
fonttools==4.59.1
gitdb==4.0.12
gitpython==3.1.45
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.9
matplotlib==3.10.5
narwhals==2.1.2
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
plotly==6.3.0
pluggy==1.6.0
protobuf==6.31.1
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
pytest-json-report==1.5.0
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
requests==2.32.4
sentry-sdk==2.34.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.2
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wandb==0.21.1
